% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleolus.R
\name{nucleolus}
\alias{nucleolus}
\title{Nucleolus solution}
\usage{
nucleolus(game, show.data = FALSE)
}
\arguments{
\item{game}{a vector that represents the cooperative game.}

\item{show.data}{logical value indicating if the function displays the 
console output (\code{TRUE}) or not (\code{FALSE}). By default the 
value is \code{FALSE}.}
}
\value{
\code{nucleolus} returns and prints the Nucleolus Solution
of associated cooperative game.
}
\description{
This function computes the nucleolus solution of a game with a maximum of 
4 agents.
}
\examples{
# Cooperative game
game <- c(68, 102, 0, 170, 710, 762, 992)
# Nucleolus solution
nucleolus(game, show.data = TRUE)

 # ------------------------
 # Nucleolus Solution
 # ------------------------
 # [1] "(149, 192, 651)"
  
}
\author{
D. Prieto
}
