\name{coopProducGame-package}
\alias{coopProducGame-package}
\alias{coopProducGame-package}
\docType{package}
\title{Cooperative aspects of linear product games}
\description{G. Owen (1975, Math. Programming 9, 358-370) assigned to each linear production process a cooperative game, a ``linear production game''. Further, he introduced a method to find a subset of the core of linear production games that verifies certain properties, which is called the ``Owen set.'' This package computes the linear production games and allocation rules associated.}

\details{
\tabular{ll}{
Package: \tab coopProductGame\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2018-07-01\cr
License: \tab GPL-3\cr
}
The most important function is \link{coopProductGame}. Other functions included in the package are auxiliary ones that can be used independently.
}
\author{
Daniel Prieto Rodríguez

Maintainer: Daniel Prieto Rodríguez <daniel.prieto.rodriguez89@gmail.com>
}
\references{

S. Cano-Berlanga, J. M. Gimenez-Gomez, and C. Vilella. Enjoying cooperative games: The r package gametheory. Working Paper No. 06; CREIP; Spain, March 2015.

D. B. Gillies. Some Theorems on n-Person Games. PhD thesis, Princeton University, 1953.

G. Owen. On the core of linear production games. Mathematical Programming, 9:358–370, 1975.

D. Schmeidler. The nucleolus of a characteristic function game. SIAM Jounal of Applied Mathematics, 17:1163–1170, 1969.

L. S. Shapley. A value for n-person games. Contributions to the theory games II, 28:124–131, 1953.

J. R. G. van Gellekom et al. Characterization of the owen set of linear production processes. Games and Economic Behavior, 32:139–156, 2000.

}
\encoding{UTF-8}

